+++ secantTests/SendTests/SendTests.swift
@@ -129,6 +129,45 @@ class SendTests: XCTestCase {
 
         await store.finish()
     }
+    
+    @MainActor func testSendFailedBeforeHitSynchronizer() async throws {
+        // the test needs to pass the exportWallet() so we simulate some in the keychain
+        try storage.importWallet(bip39: "one two three", birthday: nil)
+
+        var initialState = SendFlowReducer.State.initial
+        initialState.transactionAddressInputState = TransactionAddressTextFieldReducer.State(
+            textFieldState:
+                TCATextFieldReducer.State(
+                    validationType: nil,
+                    text: "ztestsapling1psqa06alcfj9t6s246hht3n7kcw5h900r6z40qnuu7l58qs55kzeqa98879z9hzy596dca4hmsr".redacted
+                )
+        )
+
+        let store = TestStore(
+            initialState: initialState
+        ) {
+            SendFlowReducer(networkType: .testnet)
+        }
+        
+        store.dependencies.derivationTool = .liveValue
+        store.dependencies.mainQueue = .immediate
+        store.dependencies.mnemonic = .liveValue
+        store.dependencies.sdkSynchronizer = .noOp
+        store.dependencies.walletStorage = .noOp
+        
+        let walletStorageError: ZcashError = "export failed".toZcashError()
+        store.dependencies.walletStorage.exportWallet = { throw walletStorageError }
+        
+        // simulate the sending confirmation button to be pressed
+        await store.send(.sendPressed)
+                
+        await store.receive(.sendFailed(walletStorageError)) { state in
+            state.isSending = false
+            state.alert = AlertState.sendFailure(walletStorageError)
+        }
+
+        await store.finish()
+    }
 
     func testAddressValidation_Invalid() async throws {
         let store = TestStore(